requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.0.6
django-crispy-forms==1.14.0
django-filter==22.1
psycopg2==2.9.3
python-dotenv==0.20.0
sqlparse==0.4.2
